requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
torch
torchvision
pillow
